stock/requirements.txt
Flask==3.0.2
redis==5.0.3
gunicorn==21.2.0
grpcio==1.71.0
grpcio-tools==1.71.0
pyignite==0.6.1
dotenv==0.9.9
requests==2.31.0

prometheus-client==0.20.0
prometheus-flask-exporter==0.22.4 
py-grpc-prometheus==0.8.0

redlock-py==1.0.8
